tracker: Fix the map look up
A few commits ago, an attempt to speed the map look up was done.
Unfortunatly, it was missing the case where you actually hit the speed
up (once the map is the first element of the array, you never return
it).

So, make sure that if i is 0, you return the first element of array,
while still doing the array reordering is i > 0.