capture: add simplified collector interface
This interface is intended to be used by applications that are
being profiled by Sysprof. Tooling like GTK, GStreamer, GJS, and
others can use it to get a per-thread capture writer to avoid
locking access to the writer.

Additionally, we can port the memprof collector to use this so that
we can drop the locks.