capture: simplify collector interface
This allows us to discover a few scenarios for which might require locking
among contributor threads (when a single capture file is necessary). The
common case, however, should be a multi-threaded capture writing without
locking (using a writer in TLS) which is muxed by Sysprof after the
capture has completed.