capture: use mmap() for SysprofCaptureWriter
Currently, an inferior process needs to synchronize with the observer when
profiling completes so that data is flushed to the respective
file-desciptors.  This is problematic in scenarios where the inferior either
crashes, or cannot be reliabley stopped.

Additionally, it prevents the use of per-thread capture writers since we
cannot guarantee cleanup of those threads in synchronization with
SysprofProfiler or SysprofSource::stopped().

This changes the writing to be done with mmap(), which means we need to be
graceful in the presence of trailing zeros from posix_fallocate(). Some
additional work will need to be done to make this portable to Windows and
possibly some compilation fixes for other UNIX'like OS such as FreeBSD.
Presumably for the UNIXs just mmap() flags need to be fixed.

Once this has landed, we can have lock-free access to capture writers on a
per-thread basis which helps collection of intensive operations such as
memory tracing which requires expensive backtraces.

Older versions of Sysprof will need a bit of robustness fixes so they know
to stop when they get to a frame type of 0. They already should stop there,
but in a non-graceful manner.