build: Add feature test flags to project-wide preprocessor flags
Rather than defining them piecemeal in each C and H file, and having
problems with headers being included in different orders so that
`_POSIX_C_SOURCE` gets a value which is too low, just define the feature
test flags in `meson.build`.

See `man 7 feature_test_macros`.

This is useful for when building sysprof as a subproject beneath a
project which doesn’t define these macros, or which targets an older C
standard than `gnu11`.

Signed-off-by: Philip Withnall <withnall@endlessm.com>