libsysprof-analyze: add overlay mountinfo parsing test
This just ensures that we can parse the contents of the superblock options
that we get for "overlay" mounts via toolbox/podman. These still need to
be handled from SysprofMountNamespace when translating though.

Additionally, this stuff seems very brittle for parsing, unless it is
getting escaped somewhere I'm missing currently.