libsysprof-profile: don't use FD to drive perf source
It is not enough to get writability, as that wont trigger on the perf
source. Instead, we need to check the mmap'd header and drive things off
of that.

It might be nice to eventually determine how many of the samples are from
our own process and back-off our timeout based on that.