libsysprof-profile: parse address ranges with GRegex
This uses G_REGEX_OPTIMIZE as well to get the benefit of the JIT when using
PCRE2 via GLib.

This turns out to take about half (or less) of the time that it takes to
use sscanf() to parse the template/input string. 0.03 vs 0.08 seconds on
my system.

If someone wants to take this further, and just look at the input string
rather than rely on GRegex, that'd be fine by me too.