libsysprof-gtk: fix duration render and cull text
We don't need to render text in most cases. Additionally, that requires
inflating a GObject to extract from the capture.

Instead, try to determine if we will even want to render the text before
causing any of that work to happen.

Also setup the pixel range for the layout so it may ellipsize as necessary.