libsysprof-gtk: start on chart + layers
The goal here is to be able to connect data series such as
SysprofTimeSeries or SysprofXYSeries to various chart layers. Those
chart layers get stacked in a SysprofChart.

Some glue code will be necessary to go from frames into data series but
we can experiment with how to automate that later on.